requirements.txt
# Apple Retail v2 – pipeline dependencies
# Install: pip install -r requirements.txt

# # Config
# pyyaml>=6.0

# # Spark + Delta (for ingestion and transformations)
# pyspark>=3.4.0
# delta-spark>=2.4.0

# Data quality: run_checks.py uses only PySpark (no extra deps)


absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
awsglue-dev==2021.12.30
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.40.40
botocore==1.40.50
cachetools==6.2.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.67.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
delta-spark==4.0.0
distlib==0.3.9
distro==1.9.0
Django==5.1.7
dnspython==2.7.0
entrypoints==0.4
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.48.0
google-genai==1.61.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
great_expectations==1.11.3
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kafka-python==2.2.15
keras==3.9.2
kiwisolver==1.4.8
lark==1.2.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
multidict==6.6.3
namex==0.0.9
narwhals==2.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
openai==1.76.0
opt_einsum==3.4.0
optree==0.15.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.15.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.3.3
PyMySQL==1.1.2
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-LocalAuthentication==11.0
pyobjc-framework-Security==11.0
pyOpenSSL==25.3.0
pyotp==2.8.0
pyparsing==3.2.3
pypng==0.20220715.0
pyspark==4.0.1
pytesseract==0.3.13
pytest==7.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==7.4.2
redis==7.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.19.1
s3transfer==0.14.0
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.2.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.14.0
toolz==1.1.0
tornado==6.4.2
touchid @ git+https://github.com/parafoxia/python-touch-id@b8c2c60cf8391fc4a7369faa8b8d1bf72a144f0d
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
virtualenv==20.27.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.0.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0